tests(ci): add travis-ci.org config file for unit tests
The .travis.yml file tells the continuous integration server to
build against the last 2 bugfix versions of Go, as well as the
in-development 1.3 version. It modifies the default "install:"
and "script:" commands so that docopt-go/examples isn't included.

The README.md file was also changed to show a build status badge.